refactor: TanStack Query, Zustand 적용
- tmdb 호출하는 부분은 api로 분리
- type, interface 정의는 types로 분리
- zustand store는 stores로 분리
- 로직은 hooks로 분리
